+++ packages/docs/src/pages/TextPage.tsx
@@ -1,0 +1,15 @@
+import { Flex } from '@siakit/layout'
+import { Text } from '@siakit/text'
+
+export function TextPage() {
+  return (
+    <Flex flex justify="center" align="center">
+      <Text>
+        Lorem ipsum dolor sit amet consectetur, adipisicing elit. Nesciunt
+        incidunt culpa id esse quod eius soluta, aliquam iure, impedit repellat
+        officiis, hic odit voluptates! Dolore cupiditate aspernatur sequi
+        perferendis libero?
+      </Text>
+    </Flex>
+  )
+}